Bugs 2 Spaces before "(", Tilde in Name()
Failed to process (Parameters list) if there were 2 spaces before the Open Paren '(' e.g.
Blarg Procedure   (*GROUP G)
would mangle as BLARG@F with no parms instead of @FRg.

Name('') with Pipe as used by Capesoft Reflection may have been a problem. It was confused as a line continuation. Not sure these Pipes belong in a Procedure, they are ok in Data. Fixed to not remove the Pipes. It was a Hack to avoid hard work to do it right,

Refactored some code, variable names and debug out to make it easier to read / understand.