Bug cases "Index out of Range" without Flatten
Code using String Slicing did not check for some invalid subscripts and cased runtime error on index if Flatten is not checked.
Flatten is required for multi-line prototypes so default to ON always.
Refactored some code that output debug to be better.
Refactored some code labels and comments to improve.

Window added new 'Clear' button on first tab to make easy to get rid of examples.