C5 old Pro2Exp Enhance for Longer prototypes
The C5 Pro2Exp is the original from TopSpeed so useful to check others.
It had a limit of 200 bytes so increased to 1000.
Made Window resizable so could see those 1000.
Added Paste and Copy buttons to make easier to use.
Added 'ANY' type that translates to '?' in the old
Add LONG Equate types SIGNED UNSIGNED BOOL
Fixed so took New Style MAP Procedure Prototypes that start in column 1

May have done too much since trying not to change this much.